requirements.txt
perlin_numpy @ git+https://github.com/pvigier/perlin-numpy
imageio~=2.34.1
numpy~=1.26.4
pathlib~=1.0.1
pillow~=10.3.0
matplotlib~=3.8.4